Add mockDebt function (#24)
